improved implicit source integer precision for analog FP
This shouldn't really affect much of anything outside of bugs in the emulator or other plugins or some other interference from outside the emulation process that is messing with the keyboard's state.  We further prevent signed wraparound by allowing accumulation of the analog stick coordinates up to long precision instead of just int or char precision.