degrees already type `double`, so 360 can stay an `int`.
As C operations mixing an operand of `double` type and another of `int` type compute by implicit type conversion of the `int` type one to `double`, we don't need to force the C pre-processor to parse 360 as a double-precision float.  In fact, I'd probably want to downgrade the precision to just a regular `float` in the future by merely changing the type of `degrees` without also having to muck with the 360 constant.